[darwin] Check brew update only for brew manager user

+++ envs/darwin/lib/update-local-repos.bashrc
@@ -4,10 +4,20 @@ LOCAL_REPOS=(
   $HOME/.clenv
 )
 
-tsfile="$TMPD/Homebrew.$YMD"
-if [[ $FORCE || ! -e $tsfile ]]; then
-  brew update
-  brew outdated
-  rm -f "$TMPD/Homebrew.*"
-  touch $tsfile
-fi
+check_brew_update() {
+  local tsfile="$TMPD/Homebrew.$YMD"
+  local tmpfile="/usr/local/var/homebrew/.tmp"
+  if ! touch $tmpfile &>/dev/null; then
+    echo "[info] Not the Homebrew manager. Skip."
+    return
+  fi
+  if [[ $FORCE || ! -e $tsfile ]]; then
+    brew update
+    brew outdated
+    rm -f "$TMPD/Homebrew.*"
+    touch $tsfile
+  fi
+  rm $tmpfile
+}
+
+check_brew_update
